Tests/LibWeb/Text/expected/UIEvents/mouse-boundary-events-related-target.txt

pointerover: target=left, related=null
mouseover: target=left, related=null
pointerout: target=left, related=right
mouseout: target=left, related=right
pointerover: target=right, related=left
mouseover: target=right, related=left
